nutrition-backend/requirements.txt
asgiref==3.7.2
boto3==1.34.39
botocore==1.34.39
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
httpie==3.2.2
idna==3.6
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==23.2
psycopg2-binary==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
whitenoise==6.6.0
